+++ cedar-policy-core/src/parser/unescape.rs
@@ -132,9 +132,8 @@ impl std::fmt::Display for UnescapeError {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         write!(
             f,
-            "the input `{}` is not a valid escape: {:?}",
+            "the input `{}` is not a valid escape",
             &self.input[self.range.clone()],
-            &self.err
         )
     }
 }
